Update index.js
Thank you very much for sharing this! I have three Meross surge protectors, each with three individually switched mains power outlets, and four all on/off USB power outlets, and an overall power off button. I used Charles Proxy to sniff the switching details and the channel appears to change between 1, 2, 3 for each power outlet on/off. I've found that Charles does not always report the switching, but maybe I do not understand the comms. Anyhow, I think your code could be easily made to work with more Meross products if it expected the channel value and the device attributes (Model etc.) included in the config.json also (optionally or not) - so there are no literals in the code. My local copy of your code does this (I think) but I have no idea how to nmp that to my pi. I switched my Pi on for the first time yesterday, for a couple of hours, so I'm very new to this.